refactor: split Run() into testable functions
