+++ clockify/fixtures/testclockify_projects.yaml
@@ -19,7 +19,7 @@ interactions:
       Content-Type:
       - application/json
       Date:
-      - Sat, 03 Jul 2021 21:09:44 GMT
+      - Sun, 04 Jul 2021 13:37:13 GMT
       Expected-Client-Version:
       - 1.0.0
       Expires:
@@ -57,7 +57,7 @@ interactions:
       Content-Length:
       - "0"
       Date:
-      - Sat, 03 Jul 2021 21:09:44 GMT
+      - Sun, 04 Jul 2021 13:37:13 GMT
       Expected-Client-Version:
       - 1.0.0
       Expires:
