+++ clockify/fixtures/testclockify_task.yaml
@@ -19,7 +19,7 @@ interactions:
       Content-Type:
       - application/json
       Date:
-      - Sat, 03 Jul 2021 21:09:44 GMT
+      - Sun, 04 Jul 2021 13:37:13 GMT
       Expected-Client-Version:
       - 1.0.0
       Expires:
@@ -59,7 +59,7 @@ interactions:
       Content-Type:
       - application/json
       Date:
-      - Sat, 03 Jul 2021 21:09:44 GMT
+      - Sun, 04 Jul 2021 13:37:13 GMT
       Expected-Client-Version:
       - 1.0.0
       Expires:
