+++ clockify/fixtures/testclockify_timeentries.yaml
@@ -21,7 +21,7 @@ interactions:
       Content-Type:
       - application/json
       Date:
-      - Sat, 03 Jul 2021 21:09:44 GMT
+      - Sun, 04 Jul 2021 13:37:13 GMT
       Expected-Client-Version:
       - 1.0.0
       Expires:
@@ -59,7 +59,7 @@ interactions:
       Content-Length:
       - "0"
       Date:
-      - Sat, 03 Jul 2021 21:09:44 GMT
+      - Sun, 04 Jul 2021 13:37:13 GMT
       Expected-Client-Version:
       - 1.0.0
       Expires:
@@ -99,7 +99,7 @@ interactions:
       Content-Type:
       - application/json
       Date:
-      - Sat, 03 Jul 2021 21:09:44 GMT
+      - Sun, 04 Jul 2021 13:37:14 GMT
       Expected-Client-Version:
       - 1.0.0
       Expires:
